requirements.txt
aiogram>=3.4.0
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pyrogram>=2.0.100
tgcrypto>=1.2.5
apscheduler>=3.10.4
loguru>=0.7.2
greenlet>=3.0.0
